requirements.txt
python-dotenv
langchain
langchain-openai
openai
pytest
tiktoken
pandas
scikit-learn